@@ -9,12 +9,15 @@
  * that is compatible with all Monolog API versions
  *
  * @author New Relic PHP <php-agent@newrelic.com>
+ *
+ * Updated after fork: Added support for Monolog v3.
  */
 
 namespace NewRelic\Monolog\Enricher;
 
+use DateTimeInterface;
 use Monolog\Formatter\JsonFormatter;
-use Monolog\Logger;
+use Monolog\LogRecord;
 
 /**
  * Formats record as a JSON object with transformations necessary for
@@ -27,8 +30,8 @@ abstract class AbstractFormatter extends JsonFormatter
      * @param bool $appendNewline
      */
     public function __construct(
-        $batchMode = self::BATCH_MODE_NEWLINES,
-        $appendNewline = true
+        int $batchMode = self::BATCH_MODE_NEWLINES,
+        bool $appendNewline = true,
     ) {
         // BATCH_MODE_NEWLINES is required for batch compatibility with New
         // Relic log forwarder plugins, which handle batching records. When
@@ -49,17 +52,29 @@ abstract class AbstractFormatter extends JsonFormatter
      * @param int $depth
      * @return mixed
      */
-    protected function normalize($data, $depth = 0)
+    protected function normalize($data, $depth = 0): mixed
     {
         if ($depth == 0) {
+            $data = $data instanceof LogRecord
+                ? $data->toArray()
+                : $data;
+
+            $data['timestamp'] = (int) ($data['datetime']->format('U.u') * 1000);
+
+            if ($data['datetime'] instanceof DateTimeInterface) {
+                $data['datetime'] = (array) $data['datetime'];
+            }
+
             if (isset($data['extra']['newrelic-context'])) {
-                $data = array_merge($data, $data['extra']['newrelic-context']);
+                $data = [
+                    ...$data,
+                    ...$data['extra']['newrelic-context'],
+                ];
+
                 unset($data['extra']['newrelic-context']);
             }
-            $data['timestamp'] = intval(
-                $data['datetime']->format('U.u') * 1000
-            );
         }
+
         return parent::normalize($data, $depth);
     }
 }
